Argus Refinement
Reblocked a lot of stuff, including stuff full of mimics and decorations. Did the AI as well, as somehow the previous routine disappeared. Added sonar, Made the shells supercavitators, and allowed the Argus to choose aim points underwater. Also coated the AI in HA. Cost ~125k -> ~145k, but I think it was worth it.