Rooning
I don't know what I did to the D100.

I updated the Stronzo's gun to delet more bulet per second. This was updated on the Roon as well.

I've forgotten what I've done to the Dido, I think it was mostly config. It needed to be committed so that you could move it.

The Roon's main-gun turret cap is updated, but only the one by the Stronzo gun. The Stronzo gun's barbette was fortified per se. As well, everything got configured.